requirements.txt
torch
torchvision
timm
pillow
numpy
pandas
altair
streamlit
gdown
